Share the temporary file cleanup between the two writers beside settings
The Artwork Pack warnings writer carried a copy of the settings writer's cleanup with only the error label changed. The settings helper takes the label and both writers call it.
